+++ source/app/blueprints/rest/v2/__init__.py
@@ -1,0 +1,1 @@
+api_v2_url_prefix="/api/v2/"
